+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/GlobalIllumination.cs
@@ -8,6 +8,9 @@ namespace UnityEngine.Rendering.HighDefinition
         [Tooltip("Enable ray traced global illumination.")]
         public BoolParameter rayTracing = new BoolParameter(false);
 
+        [Tooltip("Defines the layers that GI should include.")]
+        public LayerMaskParameter layerMask = new LayerMaskParameter(-1);
+
         [Tooltip("Controls the length of GI rays.")]
         public ClampedFloatParameter rayLength = new ClampedFloatParameter(10f, 0.001f, 50f);